tongue works now.
hit boxes kinda small for the fruits and the tongue itself. basically i just check if the single point is within a box. will modify tongue growth ratio/speed/other stuff later as i test it out on my phone. thinking of adding a hint button on the tree. will do that maybe